requirements.txt
Flask
tensorflow
numpy
librosa
pydub
soundfile
pywebview
scikit-learn
pyserial
pyttsx3
